Improved speed and code cleanup
