+++ Radio3.0@claudiux/files/Radio3.0@claudiux/metadata.json
@@ -1,7 +1,7 @@
 {
   "description": "The Ultimate Internet Radio Receiver & Recorder for Cinnamon",
   "max-instances": 1,
-  "version": "3.5.0",
+  "version": "3.6.0",
   "uuid": "Radio3.0@claudiux",
   "name": "Radio3.0",
   "author": "claudiux",
