+++ sound150@claudiux/files/sound150@claudiux/6.4/applet.js
@@ -49,6 +49,7 @@ const {
     source_remove,
     remove_all_sources
 } = require("./lib/mainloopTools");
+const { del_song_arts } = require("./lib/del_song_arts");
 
 const { VolumeSlider } = require("./lib/volumeSlider");
 const { ControlButton } = require("./lib/controlButton");
@@ -60,7 +61,6 @@ const EXTENSION_UUID = "OSD150@claudiux";
 const HOME_DIR = GLib.get_home_dir();
 const APPLET_DIR = HOME_DIR + "/.local/share/cinnamon/applets/" + UUID;
 const PATH2SCRIPTS = APPLET_DIR + "/scripts";
-//~ const DEL_SONG_ARTS_SCRIPT = PATH2SCRIPTS + "/del_song_arts.sh";
 
 const XDG_RUNTIME_DIR = GLib.getenv("XDG_RUNTIME_DIR");
 const TMP_ALBUMART_DIR = XDG_RUNTIME_DIR + "/AlbumArt";
@@ -1810,33 +1810,11 @@ class Sound150Applet extends Applet.TextIconApplet {
             }
         }
         if (this.title_text_old != this.title_text) {
-            //~ Util.spawnCommandLineAsync("bash -c '%s'".format(DEL_SONG_ARTS_SCRIPT));
-            this.del_song_arts();
+            del_song_arts();
         }
         this.title_text_old = this.title_text;
         //this.set_applet_label(this.title_text);
         //~ log("setAppletText: this.title_text:\n"+this.title_text, true)
-    }
-
-    del_song_arts() {
-        let paths = [XDG_RUNTIME_DIR+"/AlbumArt/song-art", XDG_RUNTIME_DIR+"/sound150/arts", XDG_RUNTIME_DIR+"/sound150/icons"];
-        for (let dir_path of paths) {
-            if (GLib.file_test(dir_path, GLib.FileTest.EXISTS)) {
-                let dir = Gio.file_new_for_path(dir_path);
-                let dir_children = dir.enumerate_children("standard::name,standard::type,standard::icon,time::modified", Gio.FileQueryInfoFlags.NONE, null);
-                var file = dir_children.next_file(null);
-                while (file != null) {
-                    let name = file.get_name();
-                    logDebug("name: " + name);
-                    if (name.startsWith("R3SongArt")) {
-                        let f = Gio.file_new_for_path(dir_path+"/"+name);
-                        f.delete(null);
-                    }
-                    file = dir_children.next_file(null);
-                }
-                dir_children.close(null);
-            }
-        }
     }
 
     _truncate(text) {
@@ -2529,8 +2507,7 @@ class Sound150Applet extends Applet.TextIconApplet {
         this._seeker = null;
         kill_playerctld();
         if (!GLib.file_test(MPV_RADIO_PID, GLib.FileTest.EXISTS)) { // Radio3.0 is not running.
-            //~ Util.spawnCommandLineAsync("bash -c '%s'".format(DEL_SONG_ARTS_SCRIPT));
-            this.del_song_arts();
+            del_song_arts();
         }
     }
