+++ sound150@claudiux/files/sound150@claudiux/6.4/lib/s150PopupMenu.js
@@ -6,7 +6,6 @@ const UUID = "sound150@claudiux";
 const HOME_DIR = GLib.get_home_dir();
 const APPLET_DIR = HOME_DIR + "/.local/share/cinnamon/applets/" + UUID;
 const PATH2SCRIPTS = APPLET_DIR + "/scripts";
-//~ const DEL_SONG_ARTS_SCRIPT = PATH2SCRIPTS + "/del_song_arts.sh";
 
 const XDG_RUNTIME_DIR = GLib.getenv("XDG_RUNTIME_DIR");
 const TMP_ALBUMART_DIR = XDG_RUNTIME_DIR + "/AlbumArt";
@@ -17,6 +16,7 @@ const ALBUMART_TITLE_FILE = TMP_ALBUMART_DIR + "/title.txt";
 const MPV_RADIO_PID = XDG_RUNTIME_DIR + "/mpv_radio_PID";
 
 const PopupMenu = imports.ui.popupMenu;
+const { del_song_arts } = require("./lib/del_song_arts");
 const { ControlButton } = require("./lib/controlButton");
 const { VolumeSlider } = require("./lib/volumeSlider");
 const Interfaces = imports.misc.interfaces;
@@ -643,8 +643,7 @@ class Player extends PopupMenu.PopupMenuSection {
         }
 
         if (old_title != this._title) {
-            //~ Util.spawnCommandLineAsync("bash -c '%s'".format(DEL_SONG_ARTS_SCRIPT));
-            this.del_song_arts();
+            del_song_arts();
             Util.spawnCommandLineAsync("bash -c %s/get_album_art.sh".format(PATH2SCRIPTS));
         }
 
@@ -743,26 +742,6 @@ class Player extends PopupMenu.PopupMenuSection {
             }
         }
         this._applet.setAppletTextIcon(this, true);
-    }
-
-    del_song_arts() {
-        let paths = [XDG_RUNTIME_DIR+"/AlbumArt/song-art", XDG_RUNTIME_DIR+"/sound150/arts", XDG_RUNTIME_DIR+"/sound150/icons"];
-        for (let dir_path of paths) {
-            if (GLib.file_test(dir_path, GLib.FileTest.EXISTS)) {
-                let dir = Gio.file_new_for_path(dir_path);
-                let dir_children = dir.enumerate_children("standard::name,standard::type,standard::icon,time::modified", Gio.FileQueryInfoFlags.NONE, null);
-                var file = dir_children.next_file(null);
-                while (file != null) {
-                    let name = file.get_name();
-                    if (name.startsWith("R3SongArt")) {
-                        let f = Gio.file_new_for_path(dir_path+"/"+name);
-                        f.delete(null);
-                    }
-                    file = dir_children.next_file(null);
-                }
-                dir_children.close(null);
-            }
-        }
     }
 
     _setStatus(status) {
