+++ sound150@claudiux/files/sound150@claudiux/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v13.2.1~20250808
+  * Code cleanup.
+
 ### v13.2.0~20250808
   * Removes and replaces the use of the `del_song_arts.sh` script.
   * Fixes [#7497](https://github.com/linuxmint/cinnamon-spices-applets/issues/7497)
